refactor(kork-secrets): Rename and document user secret cache settings
This adds documentation to the user secrets cache settings and also renames them to better fit with the existing theme of caching-related properties.
