--- kork-secrets/src/main/java/com/netflix/spinnaker/kork/secrets/SecretCacheProperties.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2022 Apple Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.spinnaker.kork.secrets;
-
-import com.netflix.spinnaker.kork.annotations.Beta;
-import com.netflix.spinnaker.kork.annotations.NonnullByDefault;
-import java.time.Duration;
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "secrets.cache")
-@Data
-@NonnullByDefault
-@Beta
-public class SecretCacheProperties {
-  private long maximumSize = 1000L;
-  private Duration expireAfterWrite = Duration.ofMinutes(45);
-  private Duration expireAfterAccess = Duration.ofMinutes(15);
-}
